refactor: check for input validity and return results
Check if there are valid simulation times in the data. If there is no
data regarding the simulation time then the cost-effectiveness
functionality does not make any sense.

Furthermore, return the computed
cost-effectiveness instead of printing it. In this way the method can
also be used as a library funtion so that the client can assign the
result to a variable

Refs: #48