fix: start orientation of the vehicle
The start orientation of the vehicle was
hard-coded before. By calculating the
Euler angles and translating them to
quaternion it works now.
NOTE: The initial direction vector must
be negative, i.e., the opposite direction.
I could not figure out why, but it seems to
work now in this way.

refs: #63